Move HostProvisionerWithCustomRealResource to separate class
